Add user creation page with modern and intuitive interface in glass morphism style
