+++ components/shared/Header/Header.jsx
@@ -123,11 +123,11 @@ function MobileNavigation(props) {
             </div>
             <nav className="mt-6">
               <ul className="-my-2 divide-y divide-zinc-100 text-base text-zinc-800 dark:divide-zinc-100/5 dark:text-zinc-300">
+                <MobileNavItem href="/">Home</MobileNavItem>
                 <MobileNavItem href="/about">About</MobileNavItem>
-                {/* <MobileNavItem href="/articles">Articles</MobileNavItem> */}
                 {/* <MobileNavItem href="/projects">Projects</MobileNavItem> */}
                 <MobileNavItem href="/auth/signin">Login</MobileNavItem>
-                <MobileNavItem href="/">Register</MobileNavItem>
+                <MobileNavItem href="/auth/signup">Register</MobileNavItem>
               </ul>
             </nav>
           </Popover.Panel>
@@ -164,11 +164,11 @@ function DesktopNavigation(props) {
   return (
     <nav {...props}>
       <ul className="flex rounded-full bg-white/90 px-3 text-sm font-medium text-zinc-800 shadow-lg shadow-zinc-800/5 ring-1 ring-zinc-900/5 backdrop-blur dark:bg-zinc-800/90 dark:text-zinc-200 dark:ring-white/10">
+        <NavItem href="/">Home</NavItem>
         <NavItem href="/about">About</NavItem>
-        {/* <NavItem href="/articles">Articles</NavItem> */}
         {/* <NavItem href="/projects">Projects</NavItem> */}
         <NavItem href="/auth/signin">Login</NavItem>
-        <NavItem href="/">Register</NavItem>
+        <NavItem href="/auth/signup">Register</NavItem>
       </ul>
     </nav>
   )
