+++ pages/auth/signup.js
@@ -1,0 +1,9 @@
+import Register from "../../components/register/register";
+
+function signup() {
+  return (
+      <Register />
+  );
+}
+
+export default signup;
